feat: admin delete

+++ src/main/java/com/privateboat/forum/backend/serviceimpl/PostServiceImpl.java
@@ -10,6 +10,7 @@ import com.privateboat.forum.backend.entity.Post;
 import com.privateboat.forum.backend.entity.UserInfo;
 import com.privateboat.forum.backend.enumerate.PostTag;
 import com.privateboat.forum.backend.enumerate.SortPolicy;
+import com.privateboat.forum.backend.enumerate.UserType;
 import com.privateboat.forum.backend.exception.PostException;
 import com.privateboat.forum.backend.repository.*;
 import com.privateboat.forum.backend.service.PostService;
@@ -214,10 +215,18 @@ public class PostServiceImpl implements PostService {
 
     @Transactional
     @Override
-    public void deletePost(Long postId) throws PostException {
+    public void deletePost(Long postId, Long userId) throws PostException {
+        Optional<UserInfo> userInfo = userInfoRepository.findByUserId(userId);
+        if (userInfo.isEmpty()) {
+            throw new PostException(PostException.PostExceptionType.VIEWER_NOT_EXIST);
+        }
         Optional<Post> post = postRepository.findByPostId(postId);
         if (post.isEmpty()) {
             throw new PostException(PostException.PostExceptionType.POST_NOT_EXIST);
+        }
+        if (post.get().getUserInfo() != userInfo.get() &&
+                userInfo.get().getUserAuth().getUserType() != UserType.ADMIN) {
+            throw new PostException(PostException.PostExceptionType.PERMISSION_DENIED);
         }
         post.get().getUserInfo().getUserStatistic().subPost();
         userStatisticRepository.save(post.get().getUserInfo().getUserStatistic());
@@ -226,10 +235,18 @@ public class PostServiceImpl implements PostService {
 
     @Transactional
     @Override
-    public void deleteComment(Long commentId) throws PostException {
+    public void deleteComment(Long commentId, Long userId) throws PostException {
+        Optional<UserInfo> userInfo = userInfoRepository.findByUserId(userId);
+        if (userInfo.isEmpty()) {
+            throw new PostException(PostException.PostExceptionType.VIEWER_NOT_EXIST);
+        }
         Optional<Comment> comment = commentRepository.findById(commentId);
         if (comment.isEmpty()) {
             throw new PostException(PostException.PostExceptionType.COMMENT_NOT_EXIST);
+        }
+        if (comment.get().getUserInfo() != userInfo.get() &&
+                userInfo.get().getUserAuth().getUserType() != UserType.ADMIN) {
+            throw new PostException(PostException.PostExceptionType.PERMISSION_DENIED);
         }
         Post post = comment.get().getPost();
         post.deleteComment(comment.get());
